server/new-backend/requirements.txt
fastapi
uvicorn
langchain
langchain-text-splitters
openai
pypdf
python-multipart
numpy
chromadb
pydantic
pydantic-settings
python-dotenv
